decouple FilesystemTweaker
