+++ phpstan.neon
@@ -17,9 +17,6 @@ parameters:
         - '#Cannot cast PhpParser\\Node\\Expr\|PhpParser\\Node\\Name to string#'
         - '#Array \(array<array<PhpParser\\Node\\Stmt>>\) does not accept array<PhpParser\\Node\\Stmt|null>#'
 
-        # array miss types
-        - '#Method Rector\\FileSystem\\FilesFinder::splitSourceToDirectoriesAndFiles\(\) should return array<array<string>\|Symfony\\Component\\Finder\\SplFileInfo> but returns array<int, array<int, string\|Symfony\\Component\\Finder\\SplFileInfo>>#'
-
         # already fixed, invalidated cache?
         - '#Access to an undefined property PhpParser\\Node\\Expr::\$args#'
