+++ src/Autoloading/AdditionalAutoloader.php
@@ -3,9 +3,9 @@
 namespace Rector\Autoloading;
 
 use Nette\Loaders\RobotLoader;
-use Nette\Utils\Strings;
 use Rector\Configuration\Option;
 use Rector\FileSystem\FileGuard;
+use Rector\Utils\FilesystemTweaker;
 use Symfony\Component\Console\Input\InputInterface;
 
 /**
@@ -29,14 +29,24 @@ final class AdditionalAutoloader
     private $autoloadDirectories = [];
 
     /**
+     * @var FilesystemTweaker
+     */
+    private $filesystemTweaker;
+
+    /**
      * @param string[] $autoloadFiles
      * @param string[] $autoloadDirectories
      */
-    public function __construct(array $autoloadFiles, array $autoloadDirectories, FileGuard $fileGuard)
-    {
+    public function __construct(
+        array $autoloadFiles,
+        array $autoloadDirectories,
+        FileGuard $fileGuard,
+        FilesystemTweaker $filesystemTweaker
+    ) {
         $this->autoloadFiles = $autoloadFiles;
         $this->autoloadDirectories = $autoloadDirectories;
         $this->fileGuard = $fileGuard;
+        $this->filesystemTweaker = $filesystemTweaker;
     }
 
     public function autoloadWithInput(InputInterface $input): void
@@ -53,20 +63,10 @@ final class AdditionalAutoloader
     {
         $this->autoloadWithInput($input);
 
-        [$files, $directories] = $this->splitSourceToDirectoriesAndFiles($source);
+        [$files, $directories] = $this->filesystemTweaker->splitSourceToDirectoriesAndFiles($source);
+        $this->autoloadFiles($files);
 
-        // @todo include the files so people don't have to do it manually?
-
-        $absoluteDirectories = [];
-
-        foreach ($directories as $directory) {
-            if (Strings::contains($directory, '*')) { // is fnmatch for directories
-                $absoluteDirectories = array_merge($absoluteDirectories, glob($directory, GLOB_ONLYDIR));
-            } else { // is classic directory
-                $absoluteDirectories[] = $directory;
-            }
-        }
-
+        $absoluteDirectories = $this->filesystemTweaker->resolveDirectoriesWithFnmatch($directories);
         if (count($absoluteDirectories)) {
             $this->autoloadDirectories($absoluteDirectories);
         }
@@ -81,27 +81,6 @@ final class AdditionalAutoloader
         }
 
         $this->autoloadFiles([$autoloadFile]);
-    }
-
-    /**
-     * @todo decouple to standalone Finder/File something
-     * @param string[] $source
-     * @return string[][]|string[]
-     */
-    private function splitSourceToDirectoriesAndFiles(array $source): array
-    {
-        $files = [];
-        $directories = [];
-
-        foreach ($source as $singleSource) {
-            if (is_file($singleSource)) {
-                $files[] = $singleSource;
-            } else {
-                $directories[] = $singleSource;
-            }
-        }
-
-        return [$files, $directories];
     }
 
     /**
